Apply bugfixes
