+++ peppy/parsers.py
@@ -87,6 +87,7 @@ class CSVTableParser(TableParser):
         """
         self.validate_path()
         self._table = pd.read_csv(self.path, **self._pandas_kwargs)
+        self._table = self._table.where(pd.notnull(self._table), None)
         return self.table
