+++ tests/data/example_peps-master/example_nextflow_config/project_config.yaml
@@ -1,0 +1,3 @@
+pep_version: "2.1.0"
+sample_table: "samplesheet.csv"
+
