+++ tests/data/example_peps-master/example_node_alias/project_config.yaml
@@ -1,0 +1,8 @@
+pep_version: "2.0.0"
+sample_table: sample_table.csv
+output_dir: $HOME/hello_looper_results
+pipeline_dir: $HOME/pipeline_dir
+
+project_modifiers:
+  import:
+    - project_config1.yaml
