+++ tests/data/example_peps-master/example_node_alias/sample_table.csv
@@ -1,0 +1,3 @@
+sample_name,protocol,file
+frog_1,anySampleType,data/frog1_data.txt
+frog_2,anySampleType,data/frog2_data.txt
