--- tests/data/example_peps-master/example_other_sample_table_index/config.yaml
@@ -1,5 +1,0 @@
-pep_version: "2.0.0"
-sample_table: "samplesheet.csv"
-subsample_table: "subsamplesheet.csv"
-sample_table_index: "sample"
-subsample_table_index: "sample"
