--- tests/data/example_peps-master/example_other_sample_table_index/samplesheet.csv
@@ -1,6 +1,0 @@
-sample,strandedness
-WT_REP1,reverse
-WT_REP2,reverse
-RAP1_UNINDUCED_REP1,reverse
-RAP1_UNINDUCED_REP2,reverse
-RAP1_IAA_30M_REP1,reverse
